+++ Libs/Optimize/CurvatureSamplingFunction.h
